update pyld and rearrange import order to work around pyld import problems
