+++ cert_schema/errors.py
@@ -1,0 +1,6 @@
+class BlockcertValidationError(Exception):
+    pass
+
+
+class InvalidUrlError(Exception):
+    pass
