+++ requirements.txt
@@ -1,5 +1,5 @@
 jsonschema>=2.6.0
-pyld>=0.7.3
 requests>=2.18.4
+pyld>=0.8.0
 tox>=2.7.0
 validators>=0.12.0
